Disable card fadeIn in android
