+++ packages/card/src/card.android.js
@@ -1,0 +1,1 @@
+export default from "./card-content";
