+++ packages/card/src/card.js
@@ -1,81 +1,13 @@
-import React, { Component } from "react";
-import { View } from "react-native";
-import Image from "@times-components/image";
+import React from "react";
 import { Animations } from "@times-components/styleguide";
 import { cardPropTypes, cardDefaultProps } from "./card-prop-types";
-import Loading from "./card-loading";
-import styles from "./styles";
+import CardContent from "./card-content";
 
-class CardComponent extends Component {
-  shouldComponentUpdate(nextProps) {
-    const { image, imageSize, isLoading } = this.props;
-    return (
-      (image && image.uri !== nextProps.image.uri) ||
-      imageSize !== nextProps.imageSize ||
-      isLoading !== nextProps.isLoading
-    );
-  }
-
-  render() {
-    const {
-      children,
-      contentContainerClass,
-      imageContainerClass,
-      image,
-      imageRatio,
-      imageSize,
-      imageStyle,
-      isLoading,
-      isReversed,
-      showImage
-    } = this.props;
-
-    const renderImage = () => {
-      if (!showImage) return null;
-
-      const imageUrl =
-        image && image.uri
-          ? `${image.uri}${imageSize ? `&resize=${imageSize}` : ``}`
-          : ``;
-
-      return (
-        <View
-          className={imageContainerClass}
-          style={[
-            styles.imageContainer,
-            imageStyle,
-            isReversed ? styles.reversedImageContainer : ""
-          ]}
-        >
-          <Image aspectRatio={imageRatio} uri={imageUrl} />
-        </View>
-      );
-    };
-
-    return (
-      <Animations.FadeIn>
-        <View
-          style={[
-            styles.cardContainer,
-            isReversed ? styles.reversedCardContainer : ""
-          ]}
-        >
-          {!isReversed ? renderImage() : null}
-          <View
-            className={contentContainerClass}
-            style={[
-              styles.contentContainer,
-              isReversed ? styles.reversedContentContainer : ""
-            ]}
-          >
-          {isLoading ? <Loading /> : children}
-          </View>
-          {isReversed ? renderImage() : null}
-        </View>
-      </Animations.FadeIn>
-    );
-  }
-}
+const CardComponent = props => (
+  <Animations.FadeIn>
+    <CardContent {...props} />
+  </Animations.FadeIn>
+);
 
 CardComponent.propTypes = cardPropTypes;
 CardComponent.defaultProps = cardDefaultProps;
